feat: responsive drawer on smaller devices
